Reuse KeyShare in client exts
