+++ src/handshake/client_hello.rs
@@ -7,6 +7,7 @@ use p256::EncodedPoint;
 use crate::buffer::*;
 use crate::config::{TlsCipherSuite, TlsConfig};
 use crate::extensions::client::{ClientExtension, PskKeyExchangeMode};
+use crate::extensions::common::{KeyShare, KeyShareEntry};
 use crate::handshake::{Random, LEGACY_VERSION};
 use crate::named_groups::NamedGroup;
 use crate::supported_versions::TLS13;
@@ -94,10 +95,10 @@ where
             }
             .encode(buf)?;
 
-            ClientExtension::KeyShare {
+            ClientExtension::KeyShare(KeyShare(KeyShareEntry {
                 group: NamedGroup::Secp256r1,
                 opaque: public_key,
-            }
+            }))
             .encode(buf)?;
 
             if let Some(server_name) = self.config.server_name {
